Rewrite the expression evaluator's literal operands
`eval/expr/literal.rs` is what an operand looks like when it is written
out rather than named: a Number, a Float, a `0z` Blob, the three string
forms, `&option` and `$ENV`.

The thing worth writing down is that both quoted forms are parsed
**twice** — once to find the closing quote and size the result, once to
fill it — and the two passes have to agree byte for byte. What keeps
them in step is a correction the measuring pass accumulates: `extra` in
`eval_string`, `reduce` in `eval_lit_string`. Each counts how much
longer or shorter the result is than the text it was read from (`\<x>`
reserves five, a doubled brace gives one back, an interpolated piece
adds the terminator it writes). Upstream's `extra` is `unsigned` and
goes on to wrap for a piece with more doubled braces than that; the sum
with the source length is what is used and stays positive either way,
so it is a plain signed counter here with the reasoning in a comment.

`eval_string`'s fill pass was a `switch` with a `FALLTHROUGH` from the
`\<` arm into the default. That is a `handled` flag now: the arms that
resolve the escape themselves set it, and `\<` that did not name a key
falls to copying the character after the backslash, exactly as the C
does. `\x`/`\u` with no hex digit after it still copies the letter on
the *next* pass of the loop rather than in the arm.

`eval_env_var` takes a `bool` (it was `int`, and its one caller passed
`evaluate as c_int`); `string2float`'s three named values are a table.
The `assert` c2rust spliced its own file and line into is a
`debug_assert!`, which is what a C `assert` compiles to.

New oracle: `lit-probe.lua` (286 cases) — every number spelling
including the `_` separators and the float shapes the reader declines,
every `\` escape with its degenerate forms, the `\<...>` family,
interpolation in both quote flavours with doubled and stray braces, 34
option spellings and the environment lookups. Byte-identical against a
pre-slice binary, as is the 59,453-case expression probe.

`test/functional/vimscript` passes under ASan with no report;
test_eval_stuff/test_expr/test_vimscript/test_let/test_listdict/
test_blob/test_float_func/test_functions are clean.

`--allow-growth`: the standing economics.

+++ crates/nvim/src/nvim/eval/expr/level.rs
@@ -721,7 +721,7 @@ pub(crate) unsafe fn eval7(
                 ret = if matches!(*(*arg).add(1) as u8, b'"' | b'\'') {
                     eval_interp_string(arg, rettv, evaluate)
                 } else {
-                    eval_env_var(arg, rettv, evaluate as c_int)
+                    eval_env_var(arg, rettv, evaluate)
                 };
             }
             b'@' => {
